Place tests in correct files
